Add ability to override configuration for the animations
