+++ js/appconfig.ts
@@ -1,4 +1,6 @@
-﻿class GameHandHistory {
+﻿import { mergeDeep } from "poker/helpers";
+
+class GameHandHistory {
     /**
      * Show game history with cards.
      */
@@ -85,37 +87,6 @@ export class AppConfig {
         allowExpandActionBlockGuestureOnlyOnMyTurn: false,
         usePortraitModeOnly: false,
     };
-}
-/**
- * Simple object check.
- * @param item
- * @returns {boolean}
- */
-function isObject(item: any) {
-    return (item && typeof item === "object" && !Array.isArray(item));
-}
-
-/**
- * Deep merge two objects.
- * @param target
- * @param ...sources
- */
-function mergeDeep(target: any, ...sources: any[]): any {
-    if (!sources.length) { return target; }
-    const source = sources.shift();
-
-    if (isObject(target) && isObject(source)) {
-        for (const key in source) {
-            if (isObject(source[key])) {
-                if (!target[key]) { Object.assign(target, { [key]: {} }); }
-                mergeDeep(target[key], source[key]);
-            } else {
-                Object.assign(target, { [key]: source[key] });
-            }
-        }
-    }
-
-    return mergeDeep(target, ...sources);
 }
 
 export type PartialConfiguration<T> = {
